Remove unused files and start to fix auth modal
