+++ packages/client/src/components/main/ModalSwitch.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { Switch } from 'react-router-dom';
 import AuthRoute from '../util/route_util';
-import AuthModal from './auth_modal/AuthModal';
 import Modal from './form_modals/Modal';
 
 export default ({ background, notice }) => {
@@ -12,28 +11,29 @@ export default ({ background, notice }) => {
         path={['/login', '/signup']}
         routeType="auth"
         notice={notice}
-        component={AuthModal}
+        modalType='auth'
+        component={Modal}
         background={background}
       />
       <AuthRoute
         exact
         path={['/shop/:shopId/edit', '/new/shop']}
         component={Modal}
-        modalType={'shop'}
+        modalType='shop'
         background={background}
       />
       <AuthRoute
         exact
         path={['/coffee/:coffeeId/edit', '/new/coffee']}
         component={Modal}
-        modalType={'coffee'}
+        modalType='coffee'
         background={background}
       />
       <AuthRoute
         exact
         path={['/relation/add', '/relation/remove']}
         component={Modal}
-        modalType={'relation'}
+        modalType='relation'
         background={background}
       />
     </Switch>
