--- packages/client/src/components/main/form_modals/CoffeeFormModal.scss
@@ -1,39 +1,0 @@
-@import 'variables';
-
-.coffee-modal-screen {
-  position: absolute;
-  display: block;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-  padding: 0 0 0 0;
-  margin: 0 0 0 0;
-  min-height: 100%;
-  background-color: rgba(26, 33, 39, 0.65);
-  z-index: 1000;
-}
-
-.coffee-modal-content {
-  background-color: #ffffff;
-  padding: 0 0 6% 0;
-  margin: auto;
-  overflow: auto;
-  margin-top: 35px;
-  display: block;
-  position: relative;
-  width: 85%;
-  height: auto;
-  max-height: 90%;
-}
-
-.coffee-modal-body {
-  margin: auto;
-  width: 80%;
-}
-
-.coffee-modal-header {
-  font-size: 20px;
-  padding: 8px 0 10px 0;
-}
-
