--- packages/client/src/components/main/form_modals/ShopFormModal.scss
@@ -1,7 +1,0 @@
-@import 'variables';
-
-.shop-modal-header {
-  font-size: 20px;
-  padding: 8px 0 10px 0;
-}
-
